requirements.txt
altair==5.1.1
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
duckdb==0.9.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.1
google-auth-oauthlib==1.1.0
gspread==5.11.2
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sql_metadata==2.9.0
sqlparse==0.4.4
st-gsheets-connection==0.0.3
streamlit==1.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
wheel==0.40.0
zipp==3.17.0
